# openapi.yml
openapi: 3.0.0
info:
  title: inventory-control API
  version: 1.0.0
paths:
  /incrementStock:
    post:
      summary: Increment product stock
      parameters:
        - name: productName
          in: query
          required: true
          schema:
            type: string
            example: "Washing Machine"
      responses:
        '200':
          description: Stock incremented successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /decrementStock:
    post:
      summary: Decrement product stock
      parameters:
        - name: productName
          in: path
          required: true
          schema:
            type: string
            example: "Washing Machine"
      responses:
        '200':
          description: Stock decremented successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /changeStock:
    post:
      summary: Change product stock
      parameters:
        - name: productName
          in: query
          required: true
          schema:
            type: string
            example: "Washing Machine"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newStock:
                  type: integer
                  example: 100
              required:
                - newStock
      responses:
        '200':
          description: Stock changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /lowStockProducts:
    get:
      summary: Get products with less than 10 items in stock
      responses:
        '200':
          description: Products with low stock
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string